+++ src/app/(main)/settings/agent/Agent.tsx
@@ -1,0 +1,31 @@
+import isEqual from 'fast-deep-equal';
+import { memo } from 'react';
+
+import { INBOX_SESSION_ID } from '@/const/session';
+import AgentSetting from '@/features/AgentSetting';
+import { useAgentStore } from '@/store/agent';
+import { agentSelectors } from '@/store/agent/selectors';
+import { useUserStore } from '@/store/user';
+import { settingsSelectors } from '@/store/user/selectors';
+
+const Agent = memo(() => {
+  const config = useAgentStore(agentSelectors.defaultAgentConfig, isEqual);
+  const meta = useUserStore(settingsSelectors.defaultAgentMeta, isEqual);
+  const [updateAgent] = useUserStore((s) => [s.updateDefaultAgent]);
+
+  return (
+    <AgentSetting
+      config={config}
+      id={INBOX_SESSION_ID}
+      meta={meta}
+      onConfigChange={(config) => {
+        updateAgent({ config });
+      }}
+      onMetaChange={(meta) => {
+        updateAgent({ meta });
+      }}
+    />
+  );
+});
+
+export default Agent;